+++ sync-for-expenses/src/codatsyncexpenses/models/shared/clientratelimitresetwebhookdata.py
@@ -35,7 +35,7 @@ class ClientRateLimitResetWebhookData:
     > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
     """
     quota_remaining: Optional[int] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('QuotaRemaining'), 'exclude': lambda f: f is ClientRateLimitResetWebhookData.UNSET }})
-    r"""Total number of request remaining for your client."""
+    r"""Total number of requests remaining for your client."""
     reset_reason: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('ResetReason'), 'exclude': lambda f: f is None }})
     r"""The reason for your rate limit quota being reset."""