+++ sync-for-expenses/src/codatsyncexpenses/models/shared/datastatus.py
@@ -8,7 +8,7 @@ from enum import Enum
 from typing import Optional
 
 class DataTypes(str, Enum):
-    r"""Available Data types"""
+    r"""Available data types"""
     ACCOUNT_TRANSACTIONS = 'accountTransactions'
     BALANCE_SHEET = 'balanceSheet'
     BANK_ACCOUNTS = 'bankAccounts'
@@ -61,7 +61,7 @@ class DataStatus:
     current_status: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('currentStatus') }})
     r"""The current status of the dataset in Codat's cache."""
     data_type: DataTypes = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('dataType') }})
-    r"""Available Data types"""
+    r"""Available data types"""
     last_successful_sync: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('lastSuccessfulSync') }})
     r"""In Codat's data model, dates and times are represented using the <a class=\\"external\\" href=\\"https://en.wikipedia.org/wiki/ISO_8601\\" target=\\"_blank\\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example: