+++ sync-for-expenses/src/codatsyncexpenses/models/shared/pushoperation.py
@@ -70,7 +70,7 @@ class PushOperation:
     > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
     """
     data_type: Optional[DataType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('dataType'), 'exclude': lambda f: f is None }})
-    r"""Available Data types"""
+    r"""Available data types"""
     error_message: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('errorMessage'), 'exclude': lambda f: f is PushOperation.UNSET }})
     r"""A message about the error."""
     timeout_in_minutes: Optional[int] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('timeoutInMinutes'), 'exclude': lambda f: f is PushOperation.UNSET }})